requirements.txt
numpy==1.26.4
faiss-cpu==1.10.0
openai==1.61.1
sentence-transformers==3.4.1
python-dotenv==1.0.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
nl_core_news_md @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_md-3.8.0/nl_core_news_md-3.8.0-py3-none-any.whl
fastapi==0.115.8
uvicorn==0.34.0
spacy==3.8.0
google-api-python-client==2.163.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
mcp==1.4.1